Add ability to use SPSCQueue in shared memory.
These changes boil down to N things.

We need some way to inform this SPSCQueue that this particular
instance may be used in shared memory.  There are lots of ways to do
this, but in order to minimize changes to the interface itself, I
chose to use the Allocator type template parameter.

Namely, if the Allocator type contains a type alias
may_be_used_in_shared_memory, then the class will enable a trivially
default constructor and a trivial destructor.

In addition, std::atomic is not an implicit lifetime type after C++20,
so the allocator can optionally contain an atomic type suitable for
use in shared memory.

Since the purpose of using a queue in shared memory is to have a
producer and consumer possibly live in separate processes, we can't
have only one slots pointer.  The reader and writer each need their
own pointer.  This did not necessitate breaking into separate
structures, but it helps, and it makes adding support for multiple
MCAST consumers much easier as well.